refactor: clean up dir structure
