+++ frontend/src/components/Routes/ProtectedInstructorRoutes.tsx
@@ -1,5 +1,5 @@
 import {Navigate, Outlet} from "react-router-dom";
-import {AuthContext} from "./AuthContext.tsx";
+import {AuthContext} from "../../context/AuthContext.ts";
 import {useContext} from "react";
 
 export default function ProtectedInstructorRoutes() {
