+++ frontend/src/pages/CourseDetails/Assignment/AssignmentCreator.tsx
@@ -1,7 +1,7 @@
-import {Assignment, AssignmentDto} from "../types/courseTypes.ts";
+import {Assignment, AssignmentDto} from "../../../types/courseTypes.ts";
 import {ChangeEvent, FormEvent, useState} from "react";
 import {useNavigate} from "react-router-dom";
-import {convertToAssignmentDtoList} from "../utils/convertToAssignmentDto.ts";
+import {convertToAssignmentDtoList} from "../../../utils/convertToAssignmentDto.ts";
 
 type AssignmentCreatorProps = {
     updateCourse: (updatedProperty: string, updatedValue: AssignmentDto[]) => void,
